requirements.txt
groq==0.18.0
instructor==1.7.2
langfuse==2.59.1
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2
